+++ elmyra.ip.access.epo/CHANGES.rst
@@ -10,6 +10,7 @@ development
 - ui: integrate 3rd-party tools via iframe (parameter "embed-item-url")
 - ui: query builder I: quick add popular attributes
 - ui: get a better place for the activity spinner
+- api/cql: automatically apply number normalization to "num" fields, too
 
 0.4.2
 =====