+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,9 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+
+0.5.0
+=====
 - ui: fix height-flickering of list entry when new drawing is lazy-loaded into carousel
 - middleware: activate caching of generated pdf documents
 - ui: make ship-mode=single-bibdata work again